Hubris solved its first lambda
We were able to fill in the placeholder for the Nat.rec's motif, the solver
is still very rough and no where near usuable yet. I think the next steps
are to do a polish pass on the whole compiler to clean up all the cruft
and re-architect it a bit, then add error reporting for unsupported cases
in the solver and merge this patch into master. This should push the
language towards usability, we also still need to add top-level matching
and match expressions.